server.js
Changed the Meetup API url to only filter for "tech" meetups versus the other options it's currently filtering for. This change is proposed so the app will return more meetups for each search as currently it returns zero meetups on average and makes it look like the app doesn't function properly.